requirements.txt
tensorflow==2.14.0
keras==2.14.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.1
matplotlib==3.7.2
seaborn==0.12.2
pillow==10.0.0
opencv-python==4.8.0.74
fastapi==0.103.0
uvicorn==0.23.2
streamlit==1.28.0
requests==2.31.0
python-multipart==0.0.6
locust==2.16.1
docker==6.1.0
pydantic==2.4.0
python-dotenv==1.0.0
gradcam==0.4.0
tf-explain==0.3.2
kagglehub==0.1.14
